+++ database/migrations/2023_09_12_185749_add_user_id_in_graduation_designation_table.php
@@ -26,8 +26,10 @@ return new class extends Migration
     public function down()
     {
         Schema::table('graduation_designation', function (Blueprint $table) {
-            $table->dropForeign(['graduation_designation_user_id_foreign']);
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('graduation_designation', 'user_id')) {
+                $table->dropForeign(['graduation_designation_user_id_foreign']);
+                $table->dropColumn('user_id');
+            }
         });
     }
 };
